update: move helper funcs
